MAINT: Rename restrictions.py to asset_restrictions.py
For clarity as to what sort of restrictions these are
