+++ zipline/finance/asset_restrictions.py
